pyproject.toml
[project]
name = "google-workspace-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dotenv>=0.9.9",
    "google-api-python-client>=2.181.0",
    "google-auth>=2.40.3",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "mcp[cli]>=1.14.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
